Extract neo concept types into separate file
